Improve handling of word boundaries, fix #146
# What changed

In addition to the minimal fix suggested by Vouillon in #146, this PR
allows `Category.inexistant` characters (i.e., `bos` and `eos`) in the
following places:

1. before `bow`
2. after `eow`
3. on both sides of `not_boundary` (but not one side alone)

# Justification of change

That minimal fix would still give an unintuitive and unconventional
interpretation of word boundaries where `bos` and `eos`...

1. are accepted on both sides of a word boundary (`bow` or `eow`)
2. are not accepted on either side of a `not_boundary`

With the minimal fix, `bow` and `eow` matches would not need to have a
word between them. E.g.,

1. in `""`, both `bow` and `eow` would match at position 0 with
   nothing between them
2. in `"#"`, `bow` would match before the `#` and `eow` after the `#`,
   even though `#` is not a word character.

In contrast, the convention in other regex libraries is that `bos` and
`eos` are not words:

1. [Ruby](https://rubular.com/r/7jdsGlU3LOPrX2)
2. [Python](https://pythex.org/?regex=%5Cb&test_string=%23&ignorecase=0&multiline=0&dotall=0&verbose=0)
3. [re2 syntax reference](https://github.com/google/re2/wiki/Syntax).
4. PCRE: can't permalink, but you can check on https://regexpal.com .
5. JavaScript: ditto, https://regexpal.com .

This PR implements this more conventional behavior.

# Use case

Jane Street is migrating from `Re2` to `Re` by writing a wrapper that
implements the `Re2` interface on top of `Re`.

In most cases where `Re2` disagrees with `Re`, we either worked around
it in the wrapper or let the wrapper diverge from real `Re2`.

This seems like a case where we can't work around it, and where `Re2`
has the more intuitive and conventional behavior.